refactor: 디트릭스 극장 레지스트리를 location.ts 로 이동
CGV·롯데시네마·메가박스가 모두 location.ts 를 쓰고 있어 파일명을 맞췄습니다.
디트릭스 응답에는 극장 좌표가 없어 좌표 기반 근접 검색 대신 극장명 매칭만
제공한다는 점을 헤더 주석에 남겼습니다.
