+++ src/services/dtryx/location.ts
@@ -1,5 +1,8 @@
 /**
- * 디트릭스 극장 카탈로그
+ * 디트릭스 극장 레지스트리 및 극장 해석 보조 모듈
+ *
+ * 다른 영화관 서비스의 `location.ts` 와 같은 자리이지만, 디트릭스 응답에는
+ * 극장 좌표가 없어 좌표 기반 근접 검색 대신 극장명 매칭만 제공합니다.
  *
  * 디트릭스는 브랜드/극장 목록을 반환하는 공개 엔드포인트를 제공하지 않습니다.
  * 그래서 조회로 확인된 극장을 상수로 관리합니다.
